Fix typos and rename the file
